refactor: migrate login page to auth structure
