+++ docs/docs/auto-docs/screens/Auth/LoginPage/LoginPage/functions/default.md
@@ -1,0 +1,13 @@
+[Admin Docs](/)
+
+***
+
+# Function: default()
+
+> **default**(): `Element`
+
+Defined in: [src/screens/Auth/LoginPage/LoginPage.tsx:53](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/Auth/LoginPage/LoginPage.tsx#L53)
+
+## Returns
+
+`Element`
