--- docs/docs/auto-docs/screens/LoginPage/LoginPage/functions/default.md
@@ -1,13 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Function: default()
-
-> **default**(): `Element`
-
-Defined in: [src/screens/LoginPage/LoginPage.tsx:53](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/LoginPage/LoginPage.tsx#L53)
-
-## Returns
-
-`Element`
